Extract EntitySoulStealOrb
