+++ config/symbols.us.stcen.txt
@@ -28,6 +28,7 @@ unk14_yVel = 0x80181148;
 unk14_startFrame = 0x80181160;
 unk14_lifetime = 0x80181164;
 g_olroxDroolCollOffsets = 0x8018116C;
+g_ESouldStelOrbSprt = 0x801811D8;
 g_Rooms = 0x801812D4;
 g_UnkPrimHelperRot = 0x8018D5F8;
 EntityBackgroundBlock = 0x8018D600;
