+++ config/symbols.us.stdre.txt
@@ -35,6 +35,7 @@ unk14_yVel = 0x8018130C;
 unk14_startFrame = 0x80181324;
 unk14_lifetime = 0x80181328;
 g_olroxDroolCollOffsets = 0x80181330;
+g_ESouldStelOrbSprt = 0x8018139C;
 g_Rooms = 0x80181498;
 g_UnkPrimHelperRot = 0x80191A5C;
 EntityUnkId11 = 0x80191A64;
